backend/requirements.txt
fastapi
openai
aiohttp
tiktoken
dotenv
opik
google-generativeai
